test(refactor): createFakeEmailClient
